Remote connections no longer requires dummy read/write connections
Specifically for this CL, `ManagedEntityStreamServer` no longer needs dummy connections.

This is possible because we've bypassed calling `Chronicle::getConnection` in the remote connections system. Now, RemoteServer's getConnection method shouldn't be called. In this CL, that method turns into an exception to trigger a failing test should this change be adverse.

Since changes have to be staged due to different owners and their approvals, this CL's change to RemoteServer paves the way for removing `RemoteServer::readConnection` and `RemoteServer::writeConnection` entirely, but will happen in a future CL.

PiperOrigin-RevId: 489066059
Change-Id: Ic58efaf7034f51dc7b2af4a86f705e99543208af