Fix flakiness in RemoteStoreIntegrationTest and make RemoteStreamServerHandler more deterministic.
Flakiness in the test was caused by the subscriptions not necessarily actually having been fully-subscribed before publishing happened combined with the fact that publishing would "complete" before necessarily being totally complete. This change makes the publish operation within RemoteStreamServerHandler continue to suspend until the server's publish method has actually completed.

The change has the side-effect of making publish call duration being dependent on how busy the server's EntityStream is (if the stream is backed-up, the publish calls will also be backed up). This is a fair tradeoff for the time being, but if we decide we'd like to address it: we could resolve this by changing the EntityStream's sharedFlow to use a buffer, or to change the buffer strategy from SUSPEND to DROP_OLDEST.

PiperOrigin-RevId: 492534273
Change-Id: Ieb982f4fd256d7c1cbd80b2e5235530a95dbd3e6