Fix ProtoDataTest for Bazel and correctly set-up dependencies for Dagger
This change scopes-down the ProtoDataTest target to only evaluate the generated DTDs and Lenses, but to no longer test the generated dagger modules. This change was made because there is [an existing issue with Kotlin and Hilt](https://github.com/bazelbuild/rules_kotlin/issues/514) which makes constructing HiltAndroidTest tests fail.  We still test this behavior internally, however.

Additionally, the chronicle_proto_data_library macro was refactored to generate the kotlin containing the DTDs and Lenses in a separate kt_android_library from the Java-generated Dagger module (which in turn is built into an `android_library` target itself - and resolves some dagger build issues previously seen).

PiperOrigin-RevId: 493342882
Change-Id: Iffb7331d8ec91ec3271fe656b575a728ef7e3eaa