Fix RemotePolicyCheckerImplTest.
It turns out that mockito-kotlin doesn't work well with inline classes, which is how kotlin.Result is implemented. This change switches from using a mock of Chronicle to using a Fake which is configured with a lambda that provides the return value of chronicle#checkPolicy(..).

Additionally, we wrap the fake with mockito-kotlin's `spy()` so that we can use argumentCaptors. However, argumentCaptors return null when initialized, so we had to switch over to using an alternative syntax for setting up the captors.

PiperOrigin-RevId: 494721853
Change-Id: I94af9a19966124c45e3cd5bf4c972fb4838b0247